refine source for mysql and tidb
